Properly handle 'automatic' flag with pkgng when installing/upgrading.
This will use the existing flag when doing upgrades to determine
whether or not to use -DINSTALLS_DEPENDS. It grabs the flag
before uninstalling the pkg.

When installing new, only the origins specified on the cmdline
will be considered non-automatic. Any dependencies brought in
will be installed with -DINSTALLS_DEPENDS

Fixes #31