If PREFIX/sbin/pkg is not available, try using the ports-mgmt/pkg pkg-static version
This fixes build on 9.1, STABLE, and CURRENT. When ports-mgmt/pkg is
deinstalled, the PKG_BIN variable will return WRKDIR/pkg-static to
use to interact with the pkgng database. This is only a temporary
measure as the normal PREFIX/sbin/pkg will be used right away after
the install phase is called.

Fixes:	#46