Change how pkg(8) is detected.
This is somewhat hacky because pkg(8) has no reliable way to
determine if it is setup and ready for use (and that the
bootstrapper is not being called). This is fixed by pkg -N,
but was not released in 9.1-R or until pkg-1.0.13. This method
invokes the bootstrapper in a safe way, and relies on 'pkg info pkg'
to determine if pkg is ready for use.

This method is also much faster than running make -V.